apps/ai-engine/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# IA & NLP
transformers==4.35.2
torch>=2.2.0
sentence-transformers==2.2.2
# chromadb==0.4.18  # Désactivé temporairement (problème de compilation)
langchain>=0.1.0

# Document Processing
PyPDF2==3.0.1
markdown==3.5.1
beautifulsoup4==4.12.2

# HTTP Client
httpx==0.25.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
numpy==1.26.2
scikit-learn==1.3.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

